qca961x: edma: cast ETH_GSTRING_LEN to size_t
In file included from include/linux/skbuff.h:17:0,
                 from include/linux/if_ether.h:23,
                 from include/uapi/linux/ethtool.h:17,
                 from include/linux/ethtool.h:16,
                 from
drivers/net/ethernet/qcom/essedma/edma_ethtool.c:16:
drivers/net/ethernet/qcom/essedma/edma_ethtool.c: In function
'edma_get_strings':
include/linux/kernel.h:712:17: warning: comparison of distinct pointer
types lacks a cast [enabled by default]
  (void) (&_min1 == &_min2);  \
                 ^
drivers/net/ethernet/qcom/essedma/edma_ethtool.c:72:5: note: in
expansion of macro 'min'
     min(ETH_GSTRING_LEN, strlen(edma_gstrings_stats[i].stat_string) +
1));
     ^

+++ include/uapi/linux/ethtool.h
@@ -523,7 +523,7 @@ struct ethtool_pauseparam {
 	__u32	tx_pause;
 };
 
-#define ETH_GSTRING_LEN		32
+#define ETH_GSTRING_LEN		((size_t)32)
 
 /**
  * enum ethtool_stringset - string set ID
